+++ src/device/tsp/Tsp/Script/IScriptEncryptor.cs
@@ -3,33 +3,21 @@ namespace cc.isr.VI.Tsp.Script;
 /// <remarks>   2025-09-19. </remarks>
 public interface IScriptEncryptor
 {
-    /// <summary>   Encrypts a binary file. </summary>
-    /// <remarks>   2025-09-19. </remarks>
-    /// <param name="sourceFile">   Source file. </param>
-    /// <param name="destination">  Destination for the. </param>
-    public void EncryptBinaryFile( string sourceFile, string destination );
+    /// <summary>   Decrypts the provided Base 64 cipher text. </summary>
+    /// <remarks>   2025-05-24. </remarks>
+    /// <param name="cipherText">   The cipher text. </param>
+    /// <returns>   A decrypted text. </returns>
+    public string DecryptFromBase64( string cipherText );
 
-    /// <summary>   Decrypt binary file. </summary>
-    /// <remarks>   2025-09-19. </remarks>
-    /// <param name="sourceFile">   Source file. </param>
-    /// <param name="destination">  Destination for the. </param>
-    public void DecryptBinaryFile( string sourceFile, string destination );
+    /// <summary>   Encrypt to base 64. </summary>
+    /// <param name="plainText">  The plain text to encrypt to Base 64 format. </param>
+    /// <returns>   The cipher text in base 64 format. </returns>
+    public string EncryptToBase64( string plainText );
 
-    /// <summary>   Loads encrypted text file. </summary>
-    /// <remarks>   2025-09-19. </remarks>
-    /// <param name="sourceFile">   Source file. </param>
-    /// <returns>   The encrypted text file. </returns>
-    public List<string> LoadEncryptedTextFile( string sourceFile );
+    /// <summary>   Query if 'contents' is encrypted. </summary>
+    /// <remarks>   2025-04-14. </remarks>
+    /// <param name="contents"> The content to be checked for encryption. </param>
+    /// <returns>   True if encrypted, false if not. </returns>
+    public bool IsEncrypted( string contents );
 
-    /// <summary>   Generates a stream from string. </summary>
-    /// <remarks>   2025-09-19. </remarks>
-    /// <param name="s">    The string. </param>
-    /// <returns>   The stream from string. </returns>
-    public Stream GenerateStreamFromString( string s );
-
-    /// <summary>   Decrypts. </summary>
-    /// <remarks>   2025-09-19. </remarks>
-    /// <param name="encrypted">    The encrypted. </param>
-    /// <returns>   A string. </returns>
-    public string Decrypt( string encrypted );
 }